refactor: move definition of Deno.build to Deno.core
