+++ cli/tests/testdata/run/internal_dynamic_import.ts
@@ -1,1 +1,1 @@
-await import("internal:runtime/01_build.js");
+await import("internal:runtime/01_errors.js");
