+++ cli/tests/testdata/run/internal_dynamic_import.ts.out
@@ -1,4 +1,4 @@
 error: Uncaught TypeError: Cannot load internal module from external code
-await import("internal:runtime/01_build.js");
+await import("internal:runtime/01_errors.js");
 ^
     at [WILDCARD]/internal_dynamic_import.ts:1:1
