+++ cli/tests/testdata/run/internal_import.ts
@@ -1,1 +1,1 @@
-import "internal:runtime/01_build.js";
+import "internal:runtime/01_errors.js";
