+++ cli/tests/testdata/run/internal_import.ts.out
@@ -1,4 +1,4 @@
-error: Unsupported scheme "internal" for module "internal:runtime/01_build.js". Supported schemes: [
+error: Unsupported scheme "internal" for module "internal:runtime/01_errors.js". Supported schemes: [
     "data",
     "blob",
     "file",
