+++ ext/node/polyfills/_process/process.ts
@@ -4,16 +4,16 @@
 // The following are all the process APIs that don't depend on the stream module
 // They have to be split this way to prevent a circular dependency
 
-import { build } from "internal:runtime/01_build.js";
+const core = globalThis.Deno.core;
 import { nextTick as _nextTick } from "internal:deno_node/_next_tick.ts";
 import { _exiting } from "internal:deno_node/_process/exiting.ts";
 import * as fs from "internal:runtime/30_fs.js";
 
 /** Returns the operating system CPU architecture for which the Deno binary was compiled */
 export function arch(): string {
-  if (build.arch == "x86_64") {
+  if (core.build.arch == "x86_64") {
     return "x64";
-  } else if (build.arch == "aarch64") {
+  } else if (core.build.arch == "aarch64") {
     return "arm64";
   } else {
     throw Error("unreachable");
